feat: implement progress poreporter for convert
