+++ crates/serde_hkx_features/src/fs.rs
@@ -79,3 +79,41 @@ where
     tracing::info!("Input: {} -> Output: {}", input.display(), output.display());
     Ok(())
 }
+
+/// Write specified or same location.
+///
+/// - `ext`: If `output` is unspecified, rewrite the `input` extension and make it the `output`. In that case, the extension.
+///
+/// # Errors
+/// - Conflict error due to simultaneous dir creation.
+/// - No write permission to the given path.
+pub fn write_sync<I, O>(
+    input: I,
+    output: Option<O>,
+    ext: &str,
+    contents: impl AsRef<[u8]>,
+) -> io::Result<()>
+where
+    I: AsRef<Path>,
+    O: AsRef<Path>,
+{
+    let input = input.as_ref();
+
+    let output = output.as_ref().map_or_else(
+        || {
+            let mut output = input.to_path_buf();
+            output.set_extension(ext);
+            Cow::Owned(output)
+        },
+        |output| Cow::Borrowed(output.as_ref()),
+    );
+
+    if let Some(parent) = output.parent() {
+        std::fs::create_dir_all(parent)?;
+    }
+    std::fs::write(&output, contents)?;
+
+    #[cfg(feature = "tracing")]
+    tracing::info!("Input: {} -> Output: {}", input.display(), output.display());
+    Ok(())
+}
