+++ crates/serde_hkx_features/src/verify/dir.rs
@@ -1,4 +1,6 @@
+use super::verify_inner;
 use crate::error::{ReproduceHkxFilesSnafu, Result};
+use crate::progress::ProgressHandler;
 use rayon::prelude::*;
 use std::path::{Path, PathBuf};
 use std::sync::atomic::{AtomicUsize, Ordering};
@@ -16,8 +18,6 @@ where
     let dir = dir.as_ref();
     verify_dir_(dir, progress)
 }
-
-use super::verify_inner;
 
 fn filter_hkx_files(dir: &Path) -> Vec<PathBuf> {
     jwalk::WalkDir::new(dir)
@@ -94,56 +94,3 @@ where
         .fail()
     }
 }
-
-/// A trait for handling progress updates during a file verification process.
-pub trait ProgressHandler {
-    /// Called when no files are found in the directory.
-    /// The default implementation does nothing.
-    fn on_empty(&self) {}
-
-    /// Sets the total number of items to process.
-    /// The default implementation does nothing.
-    ///
-    /// # Arguments
-    ///
-    /// * `total` - The total number of items to process.
-    fn on_set_total(&mut self, total: usize) {
-        let _ = total; // No-op by default
-    }
-
-    /// Increments the progress by a specified amount.
-    /// The default implementation does nothing.
-    ///
-    /// # Arguments
-    ///
-    /// * `progress` - The amount by which to increment the progress.
-    fn inc(&self, progress: u64) {
-        let _ = progress; // No-op by default
-    }
-
-    /// pat for the current progress state.
-    /// The default implementation does nothing.
-    fn on_processing_path(&self, path: &Path) {
-        let _ = path; // No-op by default
-    }
-
-    /// Called when all files have been processed.
-    /// The default implementation does nothing.
-    fn on_finish(&self) {}
-
-    /// Starts monitoring the progress with success and failure counts.
-    /// The default implementation does nothing.
-    fn start_progress_monitoring(
-        &self,
-        success_count: Arc<AtomicUsize>,
-        failure_count: Arc<AtomicUsize>,
-    ) {
-        let _ = success_count; // No-op by default
-        let _ = failure_count; // No-op by default
-    }
-}
-
-/// A default implementation of the `ProgressHandler` trait that does nothing.
-pub struct DefaultProgressMonitor;
-
-impl ProgressHandler for DefaultProgressMonitor {}
